requirements.txt
# Dependencias principales
torch>=2.0.0
torchvision>=0.15.0

# Procesamiento de datos
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Visualización
matplotlib>=3.7.0
seaborn>=0.12.0

# Métricas y evaluación
scikit-learn>=1.3.0

# Interfaz gráfica
streamlit>=1.28.0

# Utilidades
tqdm>=4.65.0

# Jupyter (opcional para notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
